AI-CICD-Platform/services/api-gateway/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0

# Authentication & Authorization
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
PyJWT>=2.6.0
pyotp>=2.8.0
python-saml>=3.0.0
authlib>=1.2.0
itsdangerous>=2.1.2

# API Gateway & Routing
starlette>=0.27.0
aiohttp>=3.8.0
aiocache>=0.12.0
backoff>=2.2.0
tenacity>=8.2.0

# Monitoring & Metrics
prometheus-client>=0.16.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-instrumentation-fastapi>=0.39.0
opentelemetry-exporter-prometheus>=1.18.0

# Rate Limiting & Circuit Breaking
aioredis>=2.0.0
limits>=3.5.0
redis>=4.5.0
cachetools>=5.3.0

# Caching
aiocache>=0.12.0
cacheout>=0.14.0
cachetools>=5.3.0
diskcache>=5.6.0

# Logging & Tracing
structlog>=23.1.0
python-json-logger>=2.0.7
opentracing>=2.4.0
jaeger-client>=4.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
httpx>=0.24.0
respx>=0.20.0
asynctest>=0.13.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0
mdx-include>=1.4.0

# Utilities
PyYAML>=6.0.1
python-multipart>=0.0.6
aiofiles>=23.1.0
jinja2>=3.1.2
ujson>=5.7.0
orjson>=3.9.0

# WebSockets
python-socketio>=5.8.0
websockets>=11.0.3
